+++ src/rc/dispatcher.rs
@@ -1,42 +1,28 @@
 use super::{
-    execute_sync_dispatcher_requests, FnsAndTraits, Listener, ListenerMap, RwLock,
-    SyncDispatcherRequest,
+    execute_dispatcher_requests,
+    Listener,
 };
-use std::{
-    hash::Hash,
-    rc::{Rc, Weak},
-};
+use std::{collections::HashMap, hash::Hash};
 
-/// In charge of !Sync dispatching to all listeners.
-/// Owns a map event-variants and
-/// [`Weak`]-references to their listeners and/or owns [`Fn`]s.
-///
-/// [`Weak`]: https://doc.rust-lang.org/std/!Sync/struct.Weak.html
-/// [`Fn`]: https://doc.rust-lang.org/std/ops/trait.Fn.html
+/// In charge of parallel dispatching to all listeners.
 pub struct Dispatcher<T>
 where
     T: PartialEq + Eq + Hash + Clone + 'static,
 {
-    events: ListenerMap<T>,
-}
-
-impl<T> Default for Dispatcher<T>
-where
-    T: PartialEq + Eq + Hash + Clone + 'static,
-{
-    fn default() -> Dispatcher<T> {
-        Dispatcher {
-            events: ListenerMap::new(),
-        }
-    }
+    events: HashMap<T, Vec<Box<dyn Listener<T> + 'static>>>,
 }
 
 impl<T> Dispatcher<T>
 where
-    T: PartialEq + Eq + Hash + Clone + 'static,
+    T: PartialEq + Eq + Hash + Clone + Sized + 'static,
 {
-    /// Adds a [`Listener`] to listen for an `event_identifier`.
-    /// If `event_identifier` is a new [`HashMap`]-key, it will be added.
+    pub fn new() -> Self {
+        Self {
+            events: HashMap::new(),
+        }
+    }
+
+    /// Adds a [`Listener`] to listen for an `event_key`.
     ///
     /// **Note**: If your `Enum` owns fields you need to consider implementing
     /// the [`Hash`]- and [`PartialEq`]-trait if you want to ignore fields,
@@ -47,26 +33,25 @@ where
     /// Adding a [`Listener`] to the dispatcher:
     ///
     /// ```rust
-    /// use hey_listen::{rc::dispatcher::Dispatcher, rc::Listener, RwLock, rc::SyncDispatcherRequest};
-    /// use std::rc::Rc;
+    /// use hey_listen::rc::{
+    ///     Listener, Dispatcher, DispatcherRequest,
+    /// };
     ///
     /// #[derive(Clone, Eq, Hash, PartialEq)]
     /// enum Event {
     ///     EventType,
     /// }
     ///
-    /// struct ListenerStruct {}
+    /// struct ListenerStruct;
     ///
     /// impl Listener<Event> for ListenerStruct {
-    ///     fn on_event(&mut self, event: &Event) -> Option<SyncDispatcherRequest> { None }
+    ///     fn on_event(&self, event: &Event) -> Option<DispatcherRequest> { None }
     /// }
     ///
-    /// fn main() {
-    ///     let listener = Rc::new(RwLock::new(ListenerStruct {}));
-    ///     let mut dispatcher: Dispatcher<Event> = Dispatcher::default();
+    /// let listener = ListenerStruct;
+    /// let mut dispatcher: Dispatcher<Event> = Dispatcher::new();
     ///
-    ///     dispatcher.add_listener(Event::EventType, &listener);
-    /// }
+    /// dispatcher.add_listener(Event::EventType, listener);
     /// ```
     ///
     /// Declaring your own [`Hash`]- and [`PartialEq`]-trait to bypass
@@ -97,130 +82,34 @@ where
     /// [`Listener`]: trait.Listener.html
     /// [`Hash`]: https://doc.rust-lang.org/std/hash/trait.Hash.html
     /// [`PartialEq`]: https://doc.rust-lang.org/std/cmp/trait.PartialEq.html
-    /// [`HashMap`]: https://doc.rust-lang.org/std/collections/struct.HashMap.html
-    pub fn add_listener<D: Listener<T> + 'static>(
+    pub fn add_listener<D: Listener<T> + Sized + 'static>(
         &mut self,
-        event_identifier: T,
-        listener: &Rc<RwLock<D>>,
+        event_key: T,
+        listener: D,
     ) {
-        if let Some(listener_collection) = self.events.get_mut(&event_identifier) {
-            listener_collection.traits.push(Rc::downgrade(
-                &(Rc::clone(listener) as Rc<RwLock<dyn Listener<T> + 'static>>),
-            ));
-
-            return;
-        }
-
-        self.events.insert(
-            event_identifier,
-            FnsAndTraits::new_with_traits(vec![Rc::downgrade(
-                &(Rc::clone(listener) as Rc<RwLock<dyn Listener<T> + 'static>>),
-            )]),
-        );
-    }
-
-    /// Adds a [`Fn`] to listen for an `event_identifier`.
-    /// If `event_identifier` is a new [`HashMap`]-key, it will be added.
-    ///
-    /// **Note**: If your `Enum` owns fields you need to consider implementing
-    /// the [`Hash`]- and [`PartialEq`]-trait if you want to ignore fields.
-    ///
-    /// # Examples
-    ///
-    /// Adding a [`Fn`] to the dispatcher:
-    ///
-    /// ```rust
-    /// use hey_listen::{rc::dispatcher::Dispatcher, rc::Listener, RwLock, rc::SyncDispatcherRequest};
-    /// use std::rc::Rc;
-    ///
-    /// #[derive(Clone, Eq, Hash, PartialEq)]
-    /// enum Event {
-    ///     EventType,
-    /// }
-    ///
-    /// struct EventListener {
-    ///     used_method: bool,
-    /// }
-    ///
-    /// impl EventListener {
-    ///     fn test_method(&mut self, _event: &Event) {
-    ///         self.used_method = true;
-    ///     }
-    /// }
-    ///
-    /// fn main() {
-    ///     let listener = Rc::new(RwLock::new(EventListener { used_method: false }));
-    ///     let mut dispatcher: Dispatcher<Event> = Dispatcher::default();
-    ///     let weak_listener_ref = Rc::downgrade(&Rc::clone(&listener));
-    ///
-    ///     let closure = Box::new(move |event: &Event| -> Option<SyncDispatcherRequest> {
-    ///         if let Some(listener) = weak_listener_ref.upgrade() {
-    ///             listener.write().test_method(&event);
-    ///
-    ///             None
-    ///         } else {
-    ///             Some(SyncDispatcherRequest::StopListening)
-    ///         }
-    ///     });
-    ///
-    ///     dispatcher.add_fn(Event::EventType, closure);
-    /// }
-    /// ```
-    ///
-    /// [`Fn`]: https://doc.rust-lang.org/std/ops/trait.Fn.html
-    /// [`Hash`]: https://doc.rust-lang.org/std/hash/trait.Hash.html
-    /// [`PartialEq`]: https://doc.rust-lang.org/std/cmp/trait.PartialEq.html
-    /// [`HashMap`]: https://doc.rust-lang.org/std/collections/struct.HashMap.html
-    pub fn add_fn(
-        &mut self,
-        event_identifier: T,
-        function: Box<dyn Fn(&T) -> Option<SyncDispatcherRequest> + 'static>,
-    ) {
-        if let Some(listener_collection) = self.events.get_mut(&event_identifier) {
-            listener_collection.fns.push(function);
-
-            return;
-        }
+        let listener = Box::new(listener);
 
         self.events
-            .insert(event_identifier, FnsAndTraits::new_with_fns(vec![function]));
+            .entry(event_key)
+            .or_insert_with(|| Vec::new())
+            .push(listener as Box<dyn Listener<T> + 'static>);
     }
 
     /// All [`Listener`]s listening to a passed `event_identifier`
     /// will be called via their implemented [`on_event`]-method.
-    /// [`Fn`]s returning [`Result`] with `Ok(())` will be retained
-    /// and `Err(SyncDispatcherRequest::StopListening)` will cause them to
-    /// be removed from the event-dispatcher.
+    /// [`Listener`]s returning an [`Option`] wrapping [`DispatcherRequest`]
+    /// with `DispatcherRequest::StopListening` will cause them
+    /// to be removed from the event-dispatcher.
     ///
     /// [`Listener`]: trait.Listener.html
     /// [`on_event`]: trait.Listener.html#tymethod.on_event
-    /// [`Error`]: enum.Error.html
-    /// [`Fn`]: https://doc.rust-lang.org/std/ops/trait.Fn.html
-    /// [`Box`]: https://doc.rust-lang.org/std/boxed/struct.Box.html
-    /// [`Result`]: https://doc.rust-lang.org/std/result/enum.Result.html
+    /// [`DispatcherRequest`]: enum.DispatcherRequest.html
+    /// [`Option`]: https://doc.rust-lang.org/std/option/enum.Option.html
     pub fn dispatch_event(&mut self, event_identifier: &T) {
-        if let Some(listener_collection) = self.events.get_mut(event_identifier) {
-            let mut found_invalid_weak_ref = false;
-
-            execute_sync_dispatcher_requests(&mut listener_collection.traits, |weak_listener| {
-                if let Some(listener) = weak_listener.upgrade() {
-                    let mut listener = listener.write();
+        if let Some(mut listener_collection) = self.events.get_mut(event_identifier) {
+            execute_dispatcher_requests(&mut listener_collection, |listener| {
                     listener.on_event(event_identifier)
-                } else {
-                    found_invalid_weak_ref = true;
-                    None
-                }
             });
-
-            execute_sync_dispatcher_requests(&mut listener_collection.fns, |callback| {
-                callback(event_identifier)
-            });
-
-            if found_invalid_weak_ref {
-                listener_collection
-                    .traits
-                    .retain(|listener| Weak::clone(listener).upgrade().is_some());
-            }
         }
     }
 }