+++ src/rc/mod.rs
@@ -1,15 +1,20 @@
-pub use super::{sync::Listener, sync::SyncDispatcherRequest};
-use std::{collections::HashMap, hash::Hash, rc::Weak};
+use std::hash::Hash;
 
-use super::RwLock;
 pub mod dispatcher;
-pub mod priority_dispatcher;
 
 pub use dispatcher::Dispatcher;
-pub use priority_dispatcher::PriorityDispatcher;
 
-type EventFunction<T> = Vec<Box<dyn Fn(&T) -> Option<SyncDispatcherRequest>>>;
-type ListenerMap<T> = HashMap<T, FnsAndTraits<T>>;
+/// Every event-receiver needs to implement this trait
+/// in order to receive dispatched events.
+/// `T` being the type you use for events, e.g. an `Enum`.
+pub trait Listener<T>
+where
+    T: PartialEq + Eq + Hash + Clone + 'static,
+{
+    /// This function will be called once a listened
+    /// event-type `T` has been dispatched.
+    fn on_event(&self, event: &T) -> Option<DispatcherRequest>;
+}
 
 /// When `execute_sync_dispatcher_requests` returns,
 /// this `enum` informs on whether the return is early
@@ -20,8 +25,26 @@ pub(crate) enum ExecuteRequestsResult {
     Stopped,
 }
 
+/// An `enum` returning a request from a listener to the single-threaded dispatcher.
+/// This `enum` is not restricted to dispatcher residing in the `sync`-module.
+/// A request will be processed by the event-dispatcher depending on the variant:
+///
+/// `StopListening` will remove your listener from the event-dispatcher.
+///
+/// `StopPropagation` will stop dispatching of the current `Event` instance.
+/// Therefore, a listener issuing this is the last receiver.
+///
+/// `StopListeningAndPropagation` a combination of first `StopListening`
+/// and then `StopPropagation`.
+#[derive(Debug)]
+pub enum DispatcherRequest {
+    StopListening,
+    StopPropagation,
+    StopListeningAndPropagation,
+}
+
 /// Iterates over the passed `vec` and applies `function` to each element.
-/// `function`'s returned [`SyncDispatcherRequest`] will instruct
+/// `function`'s returned [`SyncDispatchResult`] will instruct
 /// a procedure depending on its variant:
 ///
 /// `StopListening`: Removes item from `vec`.
@@ -36,17 +59,17 @@ pub(crate) enum ExecuteRequestsResult {
 /// inserted into `vec`.
 ///
 /// **Note**: Unlike [`retain`], `execute_sync_dispatcher_requests`
-/// can break the current iteration and is able to match [`SyncDispatcherRequest`]
+/// can break the current iteration and is able to match [`SyncDispatchResult`]
 /// and perform actions based on variants.
 ///
 /// [`retain`]: https://doc.rust-lang.org/alloc/vec/struct.Vec.html#method.retain
-/// [`SyncDispatcherRequest`]: enum.SyncDispatcherRequest.html
-pub(crate) fn execute_sync_dispatcher_requests<T, F>(
+/// [`SyncDispatchResult`]: enum.SyncDispatchResult.html
+pub(crate) fn execute_dispatcher_requests<T, F>(
     vec: &mut Vec<T>,
     mut function: F,
 ) -> ExecuteRequestsResult
 where
-    F: FnMut(&T) -> Option<SyncDispatcherRequest>,
+    F: FnMut(&T) -> Option<DispatcherRequest>,
 {
     let mut index = 0;
 
@@ -54,47 +77,19 @@ where
         if index < vec.len() {
             match function(&vec[index]) {
                 None => index += 1,
-                Some(SyncDispatcherRequest::StopListening) => {
+                Some(DispatcherRequest::StopListening) => {
                     vec.swap_remove(index);
                 }
-                Some(SyncDispatcherRequest::StopPropagation) => {
+                Some(DispatcherRequest::StopPropagation) => {
                     return ExecuteRequestsResult::Stopped
                 }
-                Some(SyncDispatcherRequest::StopListeningAndPropagation) => {
+                Some(DispatcherRequest::StopListeningAndPropagation) => {
                     vec.swap_remove(index);
                     return ExecuteRequestsResult::Stopped;
                 }
             }
         } else {
             return ExecuteRequestsResult::Finished;
-        }
-    }
-}
-
-/// Yields closures and trait-objects.
-struct FnsAndTraits<T>
-where
-    T: PartialEq + Eq + Hash + Clone + 'static,
-{
-    traits: Vec<Weak<RwLock<dyn Listener<T> + 'static>>>,
-    fns: EventFunction<T>,
-}
-
-impl<T> FnsAndTraits<T>
-where
-    T: PartialEq + Eq + Hash + Clone + 'static,
-{
-    fn new_with_traits(trait_objects: Vec<Weak<RwLock<dyn Listener<T> + 'static>>>) -> Self {
-        FnsAndTraits {
-            traits: trait_objects,
-            fns: vec![],
-        }
-    }
-
-    fn new_with_fns(fns: EventFunction<T>) -> Self {
-        FnsAndTraits {
-            traits: vec![],
-            fns,
         }
     }
 }
@@ -107,11 +102,11 @@ mod tests {
     mod execute_sync_dispatcher_requests {
         use super::*;
 
-        fn map_usize_to_request(x: &usize) -> Option<SyncDispatcherRequest> {
+        fn map_usize_to_request(x: &usize) -> Option<DispatcherRequest> {
             match *x {
-                0 => Some(SyncDispatcherRequest::StopListening),
-                1 => Some(SyncDispatcherRequest::StopPropagation),
-                2 => Some(SyncDispatcherRequest::StopListeningAndPropagation),
+                0 => Some(DispatcherRequest::StopListening),
+                1 => Some(DispatcherRequest::StopPropagation),
+                2 => Some(DispatcherRequest::StopListeningAndPropagation),
                 _ => None,
             }
         }
@@ -119,7 +114,7 @@ mod tests {
         #[test]
         fn stop_listening() {
             let mut vec = vec![0, 0, 0, 1, 1, 1, 1];
-            execute_sync_dispatcher_requests(&mut vec, map_usize_to_request);
+            execute_dispatcher_requests(&mut vec, map_usize_to_request);
 
             assert_eq!(vec, [1, 0, 0, 1, 1, 1]);
         }
@@ -127,7 +122,7 @@ mod tests {
         #[test]
         fn empty_vec() {
             let mut vec = Vec::new();
-            execute_sync_dispatcher_requests(&mut vec, map_usize_to_request);
+            execute_dispatcher_requests(&mut vec, map_usize_to_request);
 
             assert!(vec.is_empty());
         }
@@ -135,7 +130,7 @@ mod tests {
         #[test]
         fn removing_all() {
             let mut vec = vec![0, 0, 0, 0, 0, 0, 0];
-            execute_sync_dispatcher_requests(&mut vec, map_usize_to_request);
+            execute_dispatcher_requests(&mut vec, map_usize_to_request);
 
             assert!(vec.is_empty());
         }
@@ -143,7 +138,7 @@ mod tests {
         #[test]
         fn remove_one_element_and_stop() {
             let mut vec = vec![2, 0];
-            execute_sync_dispatcher_requests(&mut vec, map_usize_to_request);
+            execute_dispatcher_requests(&mut vec, map_usize_to_request);
 
             assert_eq!(vec, [0]);
         }