+++ src/sync/mod.rs
@@ -3,9 +3,14 @@ use std::hash::Hash;
 
 pub mod parallel_dispatcher;
 pub mod priority_dispatcher;
+pub mod async_dispatcher;
 
+#[cfg(feature = "parallel")]
 pub use parallel_dispatcher::ParallelDispatcher;
+#[cfg(feature = "parallel")]
 pub use priority_dispatcher::PriorityDispatcher;
+#[cfg(feature = "async")]
+pub use async_dispatcher::AsyncDispatcher;
 
 /// An `enum` returning a request from a listener to its `sync` event-dispatcher.
 /// This `enum` is not restricted to dispatcher residing in the `sync`-module.
@@ -18,6 +23,7 @@ pub use priority_dispatcher::PriorityDispatcher;
 ///
 /// `StopListeningAndPropagation` a combination of first `StopListening`
 /// and then `StopPropagation`.
+#[cfg(feature = "parallel")]
 #[derive(Debug)]
 pub enum PriorityDispatcherRequest {
     StopListening,
@@ -29,6 +35,7 @@ pub enum PriorityDispatcherRequest {
 /// this `enum` informs on whether the return is early
 /// and thus forcefully stopped or finished on its own.
 #[derive(Debug)]
+#[cfg(feature = "parallel")]
 pub(crate) enum ExecuteRequestsResult {
     Finished,
     Stopped,
@@ -37,17 +44,18 @@ pub(crate) enum ExecuteRequestsResult {
 /// Every event-receiver needs to implement this trait
 /// in order to receive dispatched events.
 /// `T` being the type you use for events, e.g. an `Enum`.
+#[cfg(feature = "parallel")]
 pub trait Listener<T>
 where
     T: PartialEq + Eq + Hash + Clone + 'static,
 {
     /// This function will be called once a listened
     /// event-type `T` has been dispatched.
-    fn on_event(&self, event: &T) -> Option<ParallelDispatcherRequest>;
+    fn on_event(&self, event: &T) -> Option<ParallelDispatchResult>;
 }
 
 /// Iterates over the passed `vec` and applies `function` to each element.
-/// `function`'s returned [`ParallelDispatcherRequest`] will instruct
+/// `function`'s returned [`ParallelDispatchResult`] will instruct
 /// a procedure depending on its variant:
 ///
 /// `StopListening`: Removes item from `vec`.
@@ -61,11 +69,12 @@ where
 /// resulting in a different order.
 ///
 /// **Note**: Unlike [`retain`], `execute_sync_dispatcher_requests`
-/// can stop the current iteration and is able to match [`ParallelDispatcherRequest`]
+/// can stop the current iteration and is able to match [`ParallelDispatchResult`]
 /// and perform actions based on variants.
 ///
 /// [`retain`]: https://doc.rust-lang.org/alloc/vec/struct.Vec.html#method.retain
-/// [`ParallelDispatcherRequest`]: enum.ParallelDispatcherRequest.html
+/// [`ParallelDispatchResult`]: enum.ParallelDispatchResult.html
+#[cfg(feature = "parallel")]
 pub(crate) fn execute_sync_dispatcher_requests<T, F>(
     vec: &mut Vec<T>,
     mut function: F,
@@ -153,18 +162,46 @@ mod tests {
 /// event-dispatcher.
 ///
 /// **Note**:
-/// Opposed to `ParallelDispatcherRequest` a [`Listener`] cannot
+/// Opposed to `ParallelDispatchResult` a [`Listener`] cannot
 /// stop propagation as the propagation is happening parallel.
 ///
 /// [`Listener`]: trait.Listener.html
+#[cfg(feature = "parallel")]
 #[derive(Debug)]
-pub enum ParallelDispatcherRequest {
+pub enum ParallelDispatchResult {
+    StopListening,
+}
+
+/// An `enum` returning a request from a [`Listener`] to its async event-dispatcher.
+///
+/// `StopListening` will remove your [`Listener`] from the
+/// event-dispatcher.
+#[derive(Debug)]
+#[cfg(feature = "async")]
+pub enum AsyncDispatchResult {
     StopListening,
 }
 
 /// Every event-receiver needs to implement this trait
 /// in order to receive dispatched events.
 /// `T` being the type you use for events, e.g. an `Enum`.
+#[cfg(feature = "async")]
+#[async_trait::async_trait]
+pub trait AsyncListener<T>
+where
+    T: PartialEq + Eq + Hash + Clone + Send + Sync + 'static,
+{
+    /// This function will be called once a listened
+    /// event-type `T` has been dispatched.
+    /// If you want to mutate the listener, consider wrapping it behind an
+    /// RwLock or Mutex.
+    async fn on_event(&self, event: &T) -> Option<AsyncDispatchResult>;
+}
+
+/// Every event-receiver needs to implement this trait
+/// in order to receive dispatched events.
+/// `T` being the type you use for events, e.g. an `Enum`.
+#[cfg(feature = "parallel")]
 pub trait ParallelListener<T>
 where
     T: PartialEq + Eq + Hash + Clone + Send + Sync + 'static,
@@ -173,12 +210,13 @@ where
     /// event-type `T` has been dispatched.
     /// If you want to mutate the listener, consider wrapping it behind an
     /// RwLock or Mutex.
-    fn on_event(&self, event: &T) -> Option<ParallelDispatcherRequest>;
+    fn on_event(&self, event: &T) -> Option<ParallelDispatchResult>;
 }
 
 /// Every event-receiver needs to implement this trait
 /// in order to receive dispatched events.
 /// `T` being the type you use for events, e.g. an `Enum`.
+#[cfg(feature = "parallel")]
 pub trait PriorityListener<T>
 where
     T: PartialEq + Eq + Hash + Clone + Send + Sync + 'static,